refactor: Move work with shiki to separate module
