+++ src/course/course-code-block.component.tsx
@@ -2,6 +2,7 @@ import { useEffect, useRef, useState } from 'react';
 import { Check, Copy } from 'lucide-react';
 
 import { cn } from '@/shadcn/lib/utils';
+import { highlightCourseCode } from './course-code-block.shiki';
 import type {
   CopyState,
   CourseCodeBlockProps,
@@ -42,100 +43,9 @@ const highlightedCodeClassName = cn(
   '[&>pre]:[-webkit-text-size-adjust:none]'
 );
 
-async function createCourseHighlighter() {
-  const [
-    { createHighlighterCore },
-    { createJavaScriptRegexEngine },
-    githubDark,
-    githubLight,
-  ] = await Promise.all([
-    import('shiki/core'),
-    import('shiki/engine/javascript'),
-    import('@shikijs/themes/github-dark'),
-    import('@shikijs/themes/github-light'),
-  ]);
-
-  return createHighlighterCore({
-    engine: createJavaScriptRegexEngine(),
-    themes: [githubDark.default, githubLight.default],
-    langs: [],
-  });
-}
-
-let highlighterPromise: ReturnType<typeof createCourseHighlighter> | null =
-  null;
-let bundledLanguagesPromise: Promise<typeof import('shiki/langs')> | null =
-  null;
-
-const languageLoadPromises = new Map<string, Promise<boolean>>();
-
-function getCourseHighlighter() {
-  highlighterPromise ??= createCourseHighlighter();
-  return highlighterPromise;
-}
-
-function getBundledLanguages() {
-  bundledLanguagesPromise ??= import('shiki/langs');
-  return bundledLanguagesPromise;
-}
-
 function normalizeLanguage(language?: string): string | null {
   const normalized = language?.trim().toLowerCase();
   return normalized || null;
-}
-
-async function ensureLanguageLoaded(language: string) {
-  const cachedPromise = languageLoadPromises.get(language);
-
-  if (cachedPromise) {
-    return cachedPromise;
-  }
-
-  const loadPromise = (async () => {
-    const [{ bundledLanguages }, highlighter] = await Promise.all([
-      getBundledLanguages(),
-      getCourseHighlighter(),
-    ]);
-    const loadLanguage =
-      bundledLanguages[language as keyof typeof bundledLanguages];
-
-    if (!loadLanguage) {
-      return false;
-    }
-
-    await highlighter.loadLanguage(loadLanguage);
-    return true;
-  })();
-
-  languageLoadPromises.set(language, loadPromise);
-  return loadPromise;
-}
-
-async function highlightCode(code: string, language: string | null) {
-  if (!language) {
-    return null;
-  }
-
-  try {
-    const isLanguageLoaded = await ensureLanguageLoaded(language);
-
-    if (!isLanguageLoaded) {
-      return null;
-    }
-
-    const highlighter = await getCourseHighlighter();
-
-    return highlighter.codeToHtml(code, {
-      lang: language,
-      themes: {
-        light: 'github-light',
-        dark: 'github-dark',
-      },
-      defaultColor: 'light',
-    });
-  } catch {
-    return null;
-  }
 }
 
 export function CourseCodeBlock({
@@ -156,7 +66,7 @@ export function CourseCodeBlock({
   useEffect(() => {
     let isMounted = true;
 
-    highlightCode(code, normalizedLanguage).then((html) => {
+    highlightCourseCode(code, normalizedLanguage).then((html) => {
       if (isMounted) {
         setHighlightResult({ key: highlightKey, html });
       }
